+++ estruturas_busca.h
@@ -3,16 +3,16 @@
 #define ESTRUTURAS_BUSCA_H
 
 typedef struct no {
-    int valor;
+    long valor;
     struct no* prox;
 } No;
 
-int busca_sequencial(int arr[], int n, int chave, int* comparacoes);
-int busca_sequencial_otimizada(int arr[], int n, int chave, int* comparacoes);
-int busca_binaria(int arr[], int n, int chave, int* comparacoes);
-No* criar_lista(int* vetor, int n, int ordenada);
-int busca_lista(No* head, int chave, int* comparacoes);
-int busca_lista_ordenada(No* head, int chave, int* comparacoes);
+long busca_sequencial(long arr[], long n, long chave, long* comparacoes);
+long busca_sequencial_otimizada(long arr[], long n, long chave, long* comparacoes);
+long busca_binaria(long arr[], long n, long chave, long* comparacoes);
+No* criar_lista(long* vetor, long n, long ordenada);
+long busca_lista(No* head, long chave, long* comparacoes);
+long busca_lista_ordenada(No* head, long chave, long* comparacoes);
 void liberar_lista(No* head);
 
 #endif
